Refactor TL codegen and remove obsolete typedefs
