+++ tl_codegen/src/error.rs
@@ -1,0 +1,9 @@
+error_chain! {
+    errors {
+        WrongTyParamsCount(ty_params: Vec<::quote::Tokens>, needed_count: usize) {
+            description("wrong number of type parameters")
+            display("wrong number of type parameters: {:?} (need {}, found {})",
+                &ty_params, needed_count, ty_params.len())
+        }
+    }
+}
